requirements.txt
customtkinter>=5.2.2
pillow>=10.0.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
psutil>=5.9.0